backend/requirements.txt
fastapi
uvicorn[standard]
fastapi-cors
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-dotenv
gspread
google-auth
cloudinary
Pillow
sqlalchemy
httpx
opencv-python
ultralytics
numpy
torch
torchvision